pyproject.toml
[tool.poetry]
name = "polarkb"
version = "0.1.0"
description = "A personal Knowledgebase system"
authors = ["jamesclair <clair.james88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
psycopg = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
